+++ src/main/java/lotto/domain/Grade.java
@@ -1,17 +1,17 @@
 package lotto.domain;
 
 public enum Grade {
-    FIRST("6개 일치 (2,000,000,000원)", 2000000000F),
-    SECOND("5개 일치, 보너스 볼 일치 (30,000,000원)", 30000000F),
-    THIRD("5개 일치 (1,500,000원)", 1500000F),
-    FOURTH("4개 일치 (50,000원)", 50000F),
-    FIFTH("3개 일치 (5,000원)", 5000F),
-    NOTHING("", 0F);
+    FIRST("6개 일치 (2,000,000,000원)", 2000000000D),
+    SECOND("5개 일치, 보너스 볼 일치 (30,000,000원)", 30000000D),
+    THIRD("5개 일치 (1,500,000원)", 1500000D),
+    FOURTH("4개 일치 (50,000원)", 50000D),
+    FIFTH("3개 일치 (5,000원)", 5000D),
+    NOTHING("", 0D);
 
     private final String message;
-    private final Float prize;
+    private final Double prize;
 
-    Grade(String message, Float prize) {
+    Grade(String message, Double prize) {
         this.message = message;
         this.prize = prize;
     }
@@ -34,7 +34,7 @@ public enum Grade {
         return message;
     }
 
-    public Float getPrize(){
+    public Double getPrize(){
         return prize;
     }
